requirements.txt
GitPython==3.1.43
Jinja2==3.1.4
MarkupSafe==2.1.5
Pygments==2.18.0
SQLAlchemy==2.0.34
aiosqlite==0.20.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.114.0
fonttools==4.53.1
gitdb==4.0.11
greenlet==3.0.3
h11==0.14.0
idna==3.8
joblib==1.4.2
jsonschema-specifications==2023.12.1
jsonschema==4.23.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
narwhals==1.6.4
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.0
psycopg2==2.9.9
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlalchemy==2.0.34
starlette==0.38.5
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
